requirements.txt
python-dotenv
openai
pyupbit
ta
selenium
webdriver-manager
pillow
youtube-transcript-api
streamlit